Make all ID on the form unique
This fixes the bug in the clear key field implementation and it is
correct way, as on a page an ele should have unique ID.

The form post by htmlx hx-post wont get impacted as htmlx will include
the nearest form only, and at server side the form lib will take the
`name` attr of the posted data to decode it into the struct.

So even there are duplicated of name however they would be in different
form thus it wont get posted.